appium-python-pytest/requirements.txt
pytest
selenium
webdriver-manager
pytest-html
pytest-xdist
Appium-Python-Client